update product quantity in cart

+++ src/main/java/com/ecommerce/MazdaCart/security/jwt/JwtUtils.java
@@ -26,7 +26,7 @@ public class JwtUtils {
 	private String jwtSecret;
 
 	@Value("${spring.jwt.expirations.tms}")
-	private int jwtExpirationTms;
+	private long jwtExpirationTms;
 
 
 	private Key getKey () {
